refactor list of restaurants component
