+++ src/App.js
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types';
 
 import AddRestaurant from './AddRestaurant';
 import HeaderPresenter from './HeaderPresenter';
-import DisplayFavoriteRestaurant from './DisplayFavoriteRestaurant';
+import DisplayFavoriteRestaurants from './DisplayFavoriteRestaurants';
 import Roulette from './Roulette';
 
 let API_KEY = 'ba91e2f7b4c7b52bef1710bd85500437';
@@ -77,27 +77,18 @@ class App extends Component {
 
 	render() {
 		return (
-				<div className="App">
-					<HeaderPresenter />
-					<AddRestaurant addFavorite={this.handleAddFavorite} />
-
-					{this.state.restaurants.map(restaurant => {
-						return (
-							<DisplayFavoriteRestaurant
-								key={restaurant.id}
-								restaurant={restaurant}
-								/>
-						);
-					})}
-
-					<Roulette
-						restaurants={this.state.restaurants}
-						resultObject={this.state.randomRestaurant}
-						menuUrl={this.state.menuUrl}
-						handleRestaurantRoulette={() => this.selectRandomRestaurant(this.state.restaurants.length)}
-						generateRandomRestaurant={() => this.generateRandomRestaurant()}
-					/>
-				</div>
+			<div className="App">
+				<HeaderPresenter />
+				<AddRestaurant addFavorite={this.handleAddFavorite} />
+				<DisplayFavoriteRestaurants restaurants={this.state.restaurants}	/>
+				<Roulette
+					restaurants={this.state.restaurants}
+					resultObject={this.state.randomRestaurant}
+					menuUrl={this.state.menuUrl}
+					handleRestaurantRoulette={() => this.selectRandomRestaurant(this.state.restaurants.length)}
+					generateRandomRestaurant={() => this.generateRandomRestaurant()}
+				/>
+			</div>
 		);
 	}
 }
